requirements.txt
# Core requirements
pandas>=1.3.0
scikit-learn>=1.0.0
numpy>=1.21.0
matplotlib>=3.5.0
joblib>=1.1.0
PyYAML>=6.0
boto3>=1.24.0

# Development/testing requirements
pytest>=7.0.0
pytest-cov>=3.0.0
pylint>=2.12.0
black>=22.3.0
mypy>=0.930